Add files via upload
This code implements an Emotion Recognition system with a graphical user interface (GUI). It captures video from a webcam and uses face detection to identify faces in real-time. The system draws rectangles around the detected faces and crops the regions of interest. These face images are then preprocessed by enhancing the contrast. A trained classifier, based on Histogram of Oriented Gradients (HOG) features, is employed to predict the emotions from the preprocessed images. The GUI updates the webcam display with the rectangles around the faces and displays the predicted emotions. Overall, this code offers an interactive platform for real-time emotion recognition, showcasing the application of computer vision techniques in analyzing facial expressions.