Add files via upload
This script trains an emotion recognition classifier using a dataset of images. The dataset directory is specified, which contains subfolders for each emotion category. The images are resized to a resolution of 64x64 pixels. The script uses the imageDatastore function to read and preprocess the images, extracting Histogram of Oriented Gradients (HOG) features. The dataset is then split into training and testing sets. The size of the feature vector is determined based on a sample image, and HOG features are computed for each image in the training and testing sets. Finally, a classifier is trained using the extracted features and saved for future use. A helper function is defined to read and preprocess images. Overall, this script enables the training of an emotion recognition classifier using HOG features and provides a reusable trained classifier for predicting emotions in new images.